feat: add uncaughtException  and unhandledRejection
